# docs/openapi.yaml
openapi: "3.0.1"
info:
  version: "1.0"
  title: Bank Cards API
  contact:
    name: kekich
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /user/send-money:
    post:
      tags:
        - user-controller
      operationId: sendMoney
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMoneyDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
...
components:
  schemas:
    SendMoneyDto:
      type: object
      properties:
        amount:
          type: integer
          format: int32
        cardNumberFrom:
          type: string
        cardNumberTo:
          type: string
...